feat: restructure authentication and player components, add new forms and context
