+++ src/app/(auth)/register/page.tsx
@@ -1,4 +1,4 @@
-import { RegistrationForm } from '#app/app/(auth)/register/registration-form'
+import { RegistrationForm } from '#app/app/(auth)/register/components/registration-form'
 import { requireAnonymous } from '#app/lib/auth.server'
 
 export default async function RegisterPage() {
