+++ src/app/games/components/add-game-form.tsx
@@ -3,7 +3,7 @@
 import { getFormProps, getTextareaProps, useForm } from '@conform-to/react'
 import { parseWithZod } from '@conform-to/zod'
 import { useActionState, useState } from 'react'
-import { useGamesContext } from '#app/app/games/lib/games-context'
+import { useGamesContext } from '#app/app/games/helpers/games-context'
 import { Checkbox, DatePicker } from '#app/components/form'
 import { Button } from '#app/components/ui/button'
 import {
