+++ src/app/players/components/player-list.tsx
