+++ src/app/statistics/components/player-stats-view.tsx
